pyproject.toml
[tool.poetry]
name = "divere"
version = "1.0.0"
description = "胶片校色工具 - 基于ACEScg Linear工作流的胶片数字化后期处理"
authors = ["V7 <vanadis@yeah.net>"]
readme = "README.md"
packages = [{include = "divere"}]
license = "MIT"
homepage = "https://github.com/your-username/DiVERE"
repository = "https://github.com/your-username/DiVERE"
keywords = ["film", "color-correction", "image-processing", "photography"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Image Processing",
]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.0"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
PyQt6 = "^6.5.0"
scipy = "^1.11.0"
imageio = "^2.31.0"
rawpy = "^0.18.0"
colour-science = "^0.4.2"
torch = "^2.0.0"
torchvision = "^0.15.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-qt = "^4.2.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
sphinx = "^7.0.0"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["divere"]
known_third_party = ["PyQt6", "numpy", "cv2", "PIL", "scipy", "imageio", "rawpy", "colour"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "PIL.*",
    "rawpy.*",
    "colour.*",
    "torch.*",
    "torchvision.*",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=divere",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gui: marks tests as GUI tests",
]

[tool.coverage.run]
source = ["divere"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 